added all the needed files from binutils. We'll build descheck starting from these sources.
I've simplified Makefile (we're not using autoconf. When descheck becomes stable and all features are implemented, we'll add autoconf support).
Added descheck.c: this will be the place where we shall add new code.

objdump segfaulted if it was compiled against the system libbfd.a, libopcodes.a and libiberty.a. If we don't identify the issue at hand we'll add the sources for these two libraries to the repository. For now we'll use some precompiled compatible versions.